simpify code
